+++ Part_2/airlinesystem/.gitignore
@@ -35,3 +35,11 @@
 /PassengerGUI$6.class
 /PassengerGUI$7.class
 /PassengerGUI.class
+/AdminGUI$23.class
+/AdminGUI$25.class
+/AdminGUI$22.class
+/AdminGUI$24.class
+/AdminGUI$26.class
+/PassengerGUI$10.class
+/PassengerGUI$8.class
+/PassengerGUI$9.class
